+++ app/src/main/AndroidManifest.xml
@@ -46,7 +46,6 @@
         </activity>
 
 
-
     </application>
 
 </manifest>
