+++ app/src/main/java/com/example/playlistmaker/presentation/searchScreen/SearchState.kt
@@ -5,7 +5,7 @@ import com.example.playlistmaker.domain.models.Track
 sealed interface SearchState {
     object ConnectionError : SearchState
     object EmptyTrackListError : SearchState
-    object TrackHistory : SearchState
+    data class TrackHistory(val tracks: List<Track>) : SearchState
     data class TrackList(val tracks: List<Track>) : SearchState
     object Loading : SearchState
     object EmptyTrackHistory : SearchState
